Move s3 tests, start cleaning tmpdir from test_parquet
